Move artifact references behind provider-neutral interface
